requirements.txt
Django>=4.2,<5.0
django-jazzmin
django-import-export
mysqlclient
pandas
numpy
matplotlib
seaborn
scikit-learn
openpyxl
xhtml2pdf
xlrd
xlwt    